--- a/pyscf/dft/numint.py
+++ b/pyscf/dft/numint.py
@@ -812,9 +812,8 @@ def _tau_dot(mol, bra, ket, wv, mask, shls_slice, ao_loc):
     mat += _dot_ao_ao(mol, bra[3], aow, mask, shls_slice, ao_loc)
     return mat
 
-def _sparse_enough(screen_index):
+def _sparse_enough(screen_index, threshold=0.5):
     # TODO: improve the turnover threshold
-    threshold = 0.5
     return numpy.count_nonzero(screen_index) < screen_index.size * threshold
 
 def _dot_ao_ao_dense(ao1, ao2, wv, out=None):
@@ -831,13 +830,8 @@ def _dot_ao_ao_dense(ao1, ao2, wv, out=None):
         return lib.ddot(ao1.T, ao2, 1, out, 1)
     else:
         assert wv.dtype == numpy.double
-        libdft.VXCdot_aow_ao_dense(
-            out.ctypes.data_as(ctypes.c_void_p),
-            ao1.ctypes.data_as(ctypes.c_void_p),
-            ao2.ctypes.data_as(ctypes.c_void_p),
-            wv.ctypes.data_as(ctypes.c_void_p),
-            ctypes.c_int(nao), ctypes.c_int(ngrids))
-    return out
+        ao1 = _scale_ao(ao1, wv.ravel())
+        return lib.ddot(ao1.T, ao2, 1, out, 1)
 
 def _dot_ao_ao_sparse(ao1, ao2, wv, nbins, screen_index, pair_mask, ao_loc,
                       hermi=0, out=None):
@@ -2856,12 +2850,32 @@ class NumInt(lib.StreamObject, LibXCMixin):
                 pair_mask = ovlp_cond < -numpy.log(self.cutoff)
             pair_mask = numpy.asarray(pair_mask, dtype=numpy.uint8)
 
+        if mo_occ is not None:
+            # eval_rho2 is more efficient unless we have a very large system
+            # for which the pair_mask is significantly sparser than the
+            # ratio of occupied to total molecular orbitals. So we use this ratio
+            # to switch between eval_rho1 and eval_rho2.
+            mo_ao_sparsity = max(0.5 * numpy.sum(mo_occ) / nao, 1e-8)
+            wts = mol.ao_loc_nr()
+            wts = (wts[1:] - wts[:-1]) / wts[-1]
+            rho1_rho2_ratio = numpy.dot(wts, pair_mask).dot(wts) / mo_ao_sparsity
+        else:
+            rho1_rho2_ratio = 0.0
+
         def make_rho(idm, ao, sindex, xctype):
-            if sindex is not None and grids is not None:
+            has_screening = sindex is not None and grids is not None
+            has_mo = mo_coeff is not None
+            if xctype == "GGA":
+                # GGA has to do more contractions using rho2 compared to rho1,
+                # so the threshold for switching to rho1 is less strict.
+                is_sparse = rho1_rho2_ratio < 1
+            else:
+                is_sparse = rho1_rho2_ratio < 0.25
+            if has_screening and (not has_mo or is_sparse):
                 return self.eval_rho1(mol, ao, dms[idm], sindex, xctype, hermi,
                                       with_lapl, cutoff=self.cutoff,
                                       ao_cutoff=grids.cutoff, pair_mask=pair_mask)
-            elif mo_coeff is not None:
+            elif has_mo:
                 return self.eval_rho2(mol, ao, mo_coeff[idm], mo_occ[idm],
                                       sindex, xctype, with_lapl)
             else: