Hi PySCF developers!

I was playing around with the DFT module and realized that there are a few parts of the DFT density and XC potential evaluation that can be improved significantly with small changes, which especially improve the performance of MGGAs.
- For small systems, MGGA calculations do not parallelize well due to the large block size in `VXCdot_aow_ao_dense`. I made the block size smaller, but I also modified `_dot_ao_ao_dense` to use `_scale_ao` and `lib.ddot` in place of `VXCdot_aow_ao_dense`, which is simpler and faster for small systems. Should I remove `VXCdot_aow_ao_dense` since it is now unused?
- For medium-size systems, the `eval_rho2` function is much faster than `eval_rho1` due to its use of molecular orbitals instead of the density matrix. I added a check to `_gen_rho_evaluator` that determines whether the speedup from sparsity (rho1) outweighs the speedup from using molecular orbitals (rho2). This especially speeds up medium-size LDA and MGGA calculations.
- For large systems, the unnecessary matrix transpose for MGGAs in `eval_rho1` actually increases cost significantly, especially when the allocated memory is low. As seen for the 300-carbon alkane below, a significant time can be spent on this transpose alone.

I benchmarked the performance of the new implementation on a number of simple organic systems and found that the new implementation is consistently faster or equivalent for the "LDA", "PBE", and "R2SCAN" functionals, with the most notable improvements being for the meta-GGA calculations. All times are in seconds. I realize this is a long table, so please let me know if you would like clarifications or details on specific systems, or if you would like additional benchmarks or code changes.

All calculations were performed with MKL as the BLAS library and native compiler optimizations turned on. Grid level 3 was used except for the 160-carbon and 300-carbon alkanes, which used grid level 1. The `max_memory` setting was 10GB, except for the case of the 300-carbon alkane, for which it was 100GB. Results were obtained on a 96-core AMD Genoa node. `Nproc` is the number of openmp processes. "ALK_Cn" is an alkane with n carbons.

**Small Systems:**
| System     | Functional | Basis         | Nproc  | T-old      | T-new      | Speedup    | 
| ---------- | ---------- | ------------- | ------ | ---------- | ---------- | ---------- |
| H2O        | LDA        | def2-qzvppd   | 1      |        0.9 |        0.9 |       0.99 | 
| H2O        | LDA        | def2-svp      | 1      |        0.3 |        0.3 |       1.04 | 
| H2O        | LDA        | def2-tzvp     | 1      |        0.3 |        0.3 |       1.01 | 
| H2O        | PBE        | def2-qzvppd   | 1      |        1.3 |        1.2 |       1.03 | 
| H2O        | PBE        | def2-svp      | 1      |        0.3 |        0.3 |       1.07 | 
| H2O        | PBE        | def2-tzvp     | 1      |        0.5 |        0.5 |       1.02 | 
| H2O        | R2SCAN     | def2-qzvppd   | 1      |        1.9 |        2.0 |       0.98 | 
| H2O        | R2SCAN     | def2-svp      | 1      |        0.4 |        0.4 |       1.08 | 
| H2O        | R2SCAN     | def2-tzvp     | 1      |        0.7 |        0.7 |       1.02 |
| C6H6       | LDA        | def2-qzvppd   | 1      |       34.0 |       34.1 |       1.00 | 
| C6H6       | LDA        | def2-qzvppd   | 16     |       10.6 |        3.2 |       3.31 | 
| C6H6       | LDA        | def2-svp      | 1      |        3.6 |        3.7 |       0.98 | 
| C6H6       | LDA        | def2-svp      | 16     |        1.4 |        0.7 |       1.90 | 
| C6H6       | LDA        | def2-tzvp     | 1      |        8.3 |        8.4 |       0.98 | 
| C6H6       | LDA        | def2-tzvp     | 16     |        3.1 |        1.1 |       2.74 | 
| C6H6       | PBE        | def2-qzvppd   | 1      |       41.6 |       41.2 |       1.01 | 
| C6H6       | PBE        | def2-qzvppd   | 16     |        3.9 |        4.0 |       0.99 | 
| C6H6       | PBE        | def2-svp      | 1      |        5.2 |        5.1 |       1.01 | 
| C6H6       | PBE        | def2-svp      | 16     |        1.0 |        1.0 |       1.05 | 
| C6H6       | PBE        | def2-tzvp     | 1      |       11.2 |       11.2 |       1.00 | 
| C6H6       | PBE        | def2-tzvp     | 16     |        1.6 |        1.5 |       1.02 | 
| C6H6       | R2SCAN     | def2-qzvppd   | 1      |       78.8 |       80.4 |       0.98 | 
| C6H6       | R2SCAN     | def2-qzvppd   | 16     |       28.9 |        6.7 |       4.33 | 
| C6H6       | R2SCAN     | def2-svp      | 1      |        7.5 |        7.7 |       0.97 | 
| C6H6       | R2SCAN     | def2-svp      | 16     |        3.1 |        1.3 |       2.42 | 
| C6H6       | R2SCAN     | def2-tzvp     | 1      |       17.7 |       18.2 |       0.97 | 
| C6H6       | R2SCAN     | def2-tzvp     | 16     |        7.9 |        2.1 |       3.80 | 

**Medium Systems:**
| System     | Functional | Basis         | Nproc  | T-old      | T-new      | Speedup    | 
| ---------- | ---------- | ------------- | ------ | ---------- | ---------- | ---------- |
| ALK_C4     | LDA        | def2-qzvppd   | 16     |       13.7 |        3.9 |       3.55 | 
| ALK_C4     | LDA        | def2-svp      | 16     |        1.6 |        0.8 |       1.91 | 
| ALK_C4     | LDA        | def2-tzvp     | 16     |        2.7 |        1.1 |       2.46 | 
| ALK_C4     | PBE        | def2-qzvppd   | 16     |        4.8 |        4.8 |       1.01 | 
| ALK_C4     | PBE        | def2-svp      | 16     |        1.1 |        1.1 |       1.02 | 
| ALK_C4     | PBE        | def2-tzvp     | 16     |        1.5 |        1.5 |       0.98 | 
| ALK_C4     | R2SCAN     | def2-qzvppd   | 16     |       35.1 |        7.8 |       4.50 | 
| ALK_C4     | R2SCAN     | def2-svp      | 16     |        3.5 |        1.4 |       2.47 | 
| ALK_C4     | R2SCAN     | def2-tzvp     | 16     |        6.7 |        2.0 |       3.33 | 
| BIPHENYL   | LDA        | def2-qzvppd   | 16     |       22.3 |       16.6 |       1.34 | 
| BIPHENYL   | LDA        | def2-svp      | 16     |        6.1 |        2.2 |       2.82 | 
| BIPHENYL   | LDA        | def2-tzvp     | 16     |       17.0 |        4.7 |       3.65 | 
| BIPHENYL   | PBE        | def2-qzvppd   | 16     |       26.9 |       22.7 |       1.19 | 
| BIPHENYL   | PBE        | def2-svp      | 16     |        3.1 |        3.0 |       1.03 | 
| BIPHENYL   | PBE        | def2-tzvp     | 16     |        6.1 |        6.0 |       1.01 | 
| BIPHENYL   | R2SCAN     | def2-qzvppd   | 16     |       60.1 |       37.4 |       1.61 | 
| BIPHENYL   | R2SCAN     | def2-svp      | 16     |       16.1 |        4.2 |       3.87 | 
| BIPHENYL   | R2SCAN     | def2-tzvp     | 16     |       47.3 |        9.7 |       4.86 | 

**Large Systems:**
| System     | Functional | Basis         | Nproc  | T-old      | T-new      | Speedup    | 
| ---------- | ---------- | ------------- | ------ | ---------- | ---------- | ---------- |
| C60        | LDA        | def2-qzvppd   | 96     |      249.9 |      220.0 |       1.14 | 
| C60        | LDA        | def2-svp      | 96     |       38.3 |       36.7 |       1.04 | 
| C60        | LDA        | def2-tzvp     | 96     |       80.7 |       68.9 |       1.17 | 
| C60        | PBE        | def2-qzvppd   | 96     |      255.6 |      245.5 |       1.04 | 
| C60        | PBE        | def2-svp      | 96     |       45.3 |       45.6 |       0.99 | 
| C60        | PBE        | def2-tzvp     | 96     |       81.6 |       81.3 |       1.00 | 
| C60        | R2SCAN     | def2-qzvppd   | 96     |      438.3 |      317.0 |       1.38 | 
| C60        | R2SCAN     | def2-svp      | 96     |       82.4 |       75.3 |       1.09 | 
| C60        | R2SCAN     | def2-tzvp     | 96     |      144.2 |      113.7 |       1.27 | 
| ALK_C160   | LDA        | def2-svp      | 96     |      171.9 |      172.4 |       1.00 | 
| ALK_C160   | LDA        | def2-tzvp     | 96     |      449.5 |      439.2 |       1.02 | 
| ALK_C160   | PBE        | def2-svp      | 96     |      199.8 |      194.8 |       1.03 | 
| ALK_C160   | PBE        | def2-tzvp     | 96     |      570.1 |      575.3 |       0.99 | 
| ALK_C160   | R2SCAN     | def2-svp      | 96     |      272.9 |      231.9 |       1.18 | 
| ALK_C160   | R2SCAN     | def2-tzvp     | 96     |     1355.3 |      905.0 |       1.50 | 
| ALK_C300   | LDA        | def2-svp      | 96     |      657.3 |      644.2 |       1.02 | 
| ALK_C300   | PBE        | def2-svp      | 96     |      674.0 |      667.1 |       1.01 | 
| ALK_C300   | R2SCAN     | def2-svp      | 96     |      923.1 |      756.5 |       1.22 |